Migrate to maven/bnd build
