--- addons/binding/org.openhab.binding.groheondus/ESH-INF/binding/binding.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<binding:binding id="groheondus" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xmlns:binding="https://openhab.org/schemas/binding/v1.0.0"
-	xsi:schemaLocation="https://openhab.org/schemas/binding/v1.0.0 https://openhab.org/schemas/binding-1.0.0.xsd">
-
-	<name>GROHE ONDUS Binding</name>
-	<description>Provides an integration for GROHE Appliances in openHAB</description>
-	<author>Florian Schmidt and Arne Wohlert</author>
-
-</binding:binding>
