--- addons/binding/org.openhab.binding.groheondus/ESH-INF/i18n/groheondus_de_DE.properties
@@ -1,11 +1,0 @@
-# binding
-binding.groheondus.name = GROHE ONDUS
-binding.groheondus.description = Stellt eine Integration mit Geräten von GROHE in openHAB zur Verfügung.
-
-# thing types
-thing-type.groheondus.senseguard.label = GROHE SENSE Guard Gerät
-thing-type.groheondus.sense.label = GROHE SENSE Gerät
-thing-type.groheondus.senseguard.description = Ein GROHE SENSE Guard
-thing-type.groheondus.sense.description = Ein GROHE SENSE
-thing-type.groheondus.account.label = GROHE ONDUS Account
-thing-type.groheondus.account.description = Dies ist die Schnittstelle zum GROHE ONDUS Account wie sie von der App verwendet wird.
