--- addons/binding/org.openhab.binding.groheondus/OSGI-INF/.gitignore
@@ -1,1 +1,0 @@
-*.xml
