--- addons/binding/org.openhab.binding.groheondus/pom.xml
@@ -1,15 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?><project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-
-  <modelVersion>4.0.0</modelVersion>
-
-  <parent>
-    <artifactId>pom</artifactId>
-    <groupId>org.openhab.binding</groupId>
-    <version>2.5.0-SNAPSHOT</version>
-  </parent>
-
-  <artifactId>org.openhab.binding.groheondus</artifactId>
-  <packaging>eclipse-plugin</packaging>
-
-  <name>GROHE ONDUS Binding</name>
-</project>
