--- addons/binding/org.openhab.binding.groheondus/src/main/java/org/openhab/binding/groheondus/internal/GroheOndusApplianceConfiguration.java
@@ -1,24 +1,0 @@
-/**
- * Copyright (c) 2010-2019 Contributors to the openHAB project
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Eclipse Public License 2.0 which is available at
- * http://www.eclipse.org/legal/epl-2.0
- *
- * SPDX-License-Identifier: EPL-2.0
- */
-package org.openhab.binding.groheondus.internal;
-
-/**
- * @author Florian Schmidt and Arne Wohlert - Initial contribution
- */
-public class GroheOndusApplianceConfiguration {
-
-    public String applianceId;
-    public int roomId;
-    public int locationId;
-    public int pollingInterval;
-}
