+++ addons/binding/pom.xml
@@ -34,7 +34,6 @@
     <module>org.openhab.binding.feed</module>
     <module>org.openhab.binding.feed.test</module>
     <module>org.openhab.binding.ftpupload</module>
-    <module>org.openhab.binding.groheondus</module>
     <module>org.openhab.binding.ihc</module>
     <module>org.openhab.binding.ihc.test</module>
     <module>org.openhab.binding.innogysmarthome</module>
