+++ bundles/org.openhab.binding.groheondus/NOTICE
@@ -1,0 +1,20 @@
+This content is produced and maintained by the openHAB project.
+
+* Project home: https://www.openhab.org
+
+== Declared Project Licenses
+
+This program and the accompanying materials are made available under the terms
+of the Eclipse Public License 2.0 which is available at
+https://www.eclipse.org/legal/epl-2.0/.
+
+== Source Code
+
+https://github.com/openhab/openhab2-addons
+
+== Third-party Content
+
+grohe-ondus-api-java
+* License: MIT License
+* Project: https://github.com/FlorianSW/grohe-ondus-api-java
+* Source: https://github.com/FlorianSW/grohe-ondus-api-java
