+++ features/karaf/openhab-addons/src/main/feature/feature.xml
@@ -278,7 +278,8 @@
 
     <feature name="openhab-binding-groheondus" description="GROHE ONDUS Binding" version="${project.version}">
         <feature>openhab-runtime-base</feature>
-        <bundle start-level="80">mvn:org.openhab.binding/org.openhab.binding.groheondus/${project.version}</bundle>
+		<bundle dependency="true">wrap:mvn:org.grohe/ondus-api/0.0.7$Bundle-Name=GROHE%20ONDUS%20Api&amp;Bundle-SymbolicName=org.grohe.ondus-api&amp;Bundle-Version=0.0.7</bundle>
+        <bundle start-level="80">mvn:org.openhab.addons.bundles/org.openhab.binding.groheondus/${project.version}</bundle>
     </feature>
 
     <feature name="openhab-binding-harmonyhub" description="Harmony Hub Binding" version="${project.version}">
